phase4/backend/requirements.txt
# Phase 3: Todo AI Chatbot Backend Dependencies

# FastAPI web framework
fastapi>=0.100.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
psycopg[binary]>=3.1.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0

# AI/Agent
openai>=2.9.0
openai-agents>=0.6.0
litellm>=1.40.0,<1.57.5

# Settings
python-dotenv>=1.0.0
pydantic-settings>=2.2.1

# HTTP client
httpx>=0.27.0
